+++ src/meico/mpm/elements/maps/RubatoMap.java
@@ -71,14 +71,7 @@ public class RubatoMap extends GenericMap {
      */
     protected void parseData(Element xml) throws InvalidDataException {
         super.parseData(xml);
-
         this.setType("rubatoMap");            // make sure this is really a "rubatoMap"
-
-        // make sure it has a startStyle attribute
-        Attribute startStyleAtt = Helper.getAttribute("startStyle", this.getXml());
-        if (startStyleAtt == null) {
-            this.getXml().addAttribute(new Attribute("startStyle", ""));
-        }
     }
 
     /**
@@ -249,7 +242,6 @@ public class RubatoMap extends GenericMap {
                 rd.xmlId = att.getValue();
 
             // get the style that applies to this date
-            rd.styleName = Helper.getAttributeValue("startStyle", this.getXml());
             for (int j = index; j >= 0; --j) {                                              // find the first style switch at or before date
                 Element s = this.elements.get(j).getValue();
                 if (s.getLocalName().equals("style")) {